added delete function and several supporting methods
POSTMORTEM:
My delete function(rightfully) looks like trash compared to other's
solutions. With my unfamiliarity with binary search trees it didn't occur
to me to swap a node at the bottom for a node at the top. I ended up
recursivly going down the tree and shifting each node by 1(or 2 in an edge case).
Because of this, my method is exponentailly more complicated than it
needed to be and has about 1 billion checks for edge cases. Regardless,
I'll keep it as is and learn better going forward.